mcp-servers/device-controller/requirements.txt
# Tuya MCP SDK
# Install from: git clone https://github.com/tuya/tuya-mcp-sdk.git
# Then: cd tuya-mcp-sdk/mcp-python && pip install -e .

# MCP Library (for local tools server)
mcp>=1.0.0

# Other dependencies
python-dotenv>=1.0.0
